+++ templates/src/app.js
@@ -1,68 +1,20 @@
-var aws = require("aws-sdk");
-var cfsign = require("aws-cloudfront-sign");
 var dotenv = require("dotenv");
 var express = require("express");
 var morgan = require("morgan");
 
 var { connect } = require("./db");
-<%if eq (index .Params `userAuth`) "yes" %>var userAuth = require("./app/auth");
+const statusRoutes = require("./app/status");
+<%if eq (index .Params `fileUploads`) "yes" %>const fileUploadRoutes = require("./app/fileupload");
+<% end %><%if eq (index .Params `userAuth`) "yes" %>const authRoutes = require("./app/auth");
 <% end %>
 dotenv.config();
 var app = express();
 app.use(morgan("combined"));
-var s3 = new aws.S3();
-<%if eq (index .Params `userAuth`) "yes" %>
-app.use("/", userAuth);
+
+app.use("/status", statusRoutes);
+<%if eq (index .Params `userAuth`) "yes" %>app.use("/", authRoutes);
+<% end %><%if eq (index .Params `fileUploads`) "yes" %>app.use("/", fileUploadRoutes);
 <% end %>
-app.get("/presigned/:key", (req, res) => {
-  var params = {
-    Bucket: process.env.S3_BUCKET,
-    Fields: {
-      key: req.params.key,
-    },
-  };
-
-  s3.createPresignedPost(params, (err, data) => {
-    if (err) {
-      console.error(err);
-      res.sendStatus(500);
-    } else {
-      console.log(data);
-      res.send(data);
-    }
-  });
-});
-
-app.get("/:key", (req, res) => {
-  var params = {
-    keypairId: process.env.CF_KEYPAIR_ID,
-    privateKeyString: process.env.CF_KEYPAIR_SECRET_KEY,
-    expireTime: new Date().getTime() + 30000, // defaults to 30s
-  };
-
-  var url = cfsign.getSignedUrl(
-    `https://files.${process.env.DOMAIN}/${req.params.key}`,
-    params
-  );
-
-  console.log(url);
-  res.redirect(url);
-});
-
-app.get("/status/ready", (req, res) => {
-  res.send("OK");
-});
-
-app.get("/status/alive", (req, res) => {
-  res.send("OK");
-});
-
-app.get("/status/about", (req, res) => {
-  res.send({
-    podName: process.env.POD_NAME,
-  });
-});
-
 var port = process.env.SERVER_PORT;
 if (!port) {
   port = 3000;
